phase1/requirements.txt
# Phase 1 - Dépendances avec Kafka
# Installation: pip install -r requirements.txt

# LLM Framework
anthropic>=0.40.0

# Data Validation
pydantic>=2.5.0

# Environment
python-dotenv>=1.0.0

# Kafka
confluent-kafka>=2.3.0
